Refactor according to wso2-extensions to siddhi-io repo move
