+++ README.md
@@ -27,7 +27,7 @@ Latest API Docs is <a target="_blank" href="https://wso2-extensions.github.io/si
 
 ```
      <dependency>
-        <groupId>org.wso2.extension.siddhi.map.text</groupId>
+        <groupId>io.siddhi.extension.map.text</groupId>
         <artifactId>siddhi-map-text</artifactId>
         <version>x.x.x</version>
      </dependency>
