+++ component/pom.xml
@@ -20,7 +20,7 @@
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <groupId>org.wso2.extension.siddhi.map.text</groupId>
+        <groupId>io.siddhi.extension.map.text</groupId>
         <artifactId>siddhi-map-text-parent</artifactId>
         <version>2.0.1-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
@@ -110,7 +110,7 @@
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Export-Package>
-                            org.wso2.extension.siddhi.map.text.*
+                            io.siddhi.extension.map.text*
                         </Export-Package>
                         <Import-Package>
                             org.wso2.siddhi.core.*;version="${siddhi.version.range}",
