+++ component/src/main/java/io/siddhi/extension/map/text/sinkmapper/TextSinkMapper.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.map.text.sinkmapper;
+package io.siddhi.extension.map.text.sinkmapper;
 
 import com.github.mustachejava.DefaultMustacheFactory;
 import com.github.mustachejava.Mustache;
