+++ component/src/main/java/io/siddhi/extension/map/text/sourcemapper/TextSourceMapper.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.extension.siddhi.map.text.sourcemapper;
+package io.siddhi.extension.map.text.sourcemapper;
 
 import io.siddhi.annotation.Example;
 import io.siddhi.annotation.Extension;
