+++ component/src/test/java/io/siddhi/extension/map/text/sinkmapper/TextDefaultSinkMapperTestCase.java
@@ -16,16 +16,16 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.map.text.sinkmapper;
+package io.siddhi.extension.map.text.sinkmapper;
 
 import io.siddhi.core.SiddhiAppRuntime;
 import io.siddhi.core.SiddhiManager;
 import io.siddhi.core.event.Event;
-import io.siddhi.core.exception.SiddhiAppCreationException;
 import io.siddhi.core.stream.input.InputHandler;
 import io.siddhi.core.stream.output.sink.InMemorySink;
 import io.siddhi.core.util.SiddhiTestHelper;
 import io.siddhi.core.util.transport.InMemoryBroker;
+import io.siddhi.extension.map.text.sinkmapper.util.HttpServerListenerHandler;
 import io.siddhi.query.api.SiddhiApp;
 import io.siddhi.query.api.annotation.Annotation;
 import io.siddhi.query.api.definition.Attribute;
@@ -41,7 +41,6 @@ import org.testng.AssertJUnit;
 import org.testng.TestException;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import org.wso2.extension.siddhi.map.text.sinkmapper.util.HttpServerListenerHandler;
 
 import java.io.File;
 import java.io.IOException;
@@ -52,8 +51,8 @@ import java.util.concurrent.atomic.AtomicInteger;
 /**
  * Test Case for QLT type output text mapper.
  */
-public class TextCustomSinkMapperTestCase {
-    private static final Logger log = Logger.getLogger(TextCustomSinkMapperTestCase.class);
+public class TextDefaultSinkMapperTestCase {
+    private static final Logger log = Logger.getLogger(TextDefaultSinkMapperTestCase.class);
     private AtomicInteger wso2Count = new AtomicInteger(0);
     private AtomicInteger ibmCount = new AtomicInteger(0);
     private int waitTime = 50;
@@ -66,223 +65,8 @@ public class TextCustomSinkMapperTestCase {
     }
 
     @Test
-    public void testTextSinkCustomMapping() throws InterruptedException {
-        log.info("Test custom text mapping.");
-        List<Object> onMessageList = new ArrayList<Object>();
-
-        InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
-            @Override
-            public void onMessage(Object msg) {
-                wso2Count.incrementAndGet();
-                onMessageList.add(msg);
-            }
-
-            @Override
-            public String getTopic() {
-                return "WSO2";
-            }
-        };
-
-        InMemoryBroker.Subscriber subscriberIBM = new InMemoryBroker.Subscriber() {
-            @Override
-            public void onMessage(Object msg) {
-                ibmCount.incrementAndGet();
-                onMessageList.add(msg);
-            }
-
-            @Override
-            public String getTopic() {
-                return "IBM";
-            }
-        };
-
-        //subscribe to "inMemory" broker per topic
-        InMemoryBroker.subscribe(subscriberWSO2);
-        InMemoryBroker.subscribe(subscriberIBM);
-
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text', @payload(\"Stock price of {{symbol}} " +
-                "is {{price}}\"))) " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        siddhiManager.setExtension("sink:inMemory", InMemorySink.class);
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
-
-        siddhiAppRuntime.start();
-        stockStream.send(new Object[]{"WSO2", 55.6f, 100L});
-        stockStream.send(new Object[]{"IBM", 75.6f, 100L});
-        stockStream.send(new Object[]{"WSO2", 57.6f, 100L});
-        SiddhiTestHelper.waitForEvents(waitTime, 2, wso2Count, timeout);
-        SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
-
-        //assert event count
-        Assert.assertEquals(wso2Count.get(), 2);
-        Assert.assertEquals(ibmCount.get(), 1);
-        //assert custom text
-        Assert.assertEquals(onMessageList.get(0).toString(), "Stock price of WSO2 is 55.6");
-        Assert.assertEquals(onMessageList.get(1).toString(), "Stock price of IBM is 75.6");
-        Assert.assertEquals(onMessageList.get(2).toString(), "Stock price of WSO2 is 57.6");
-        siddhiAppRuntime.shutdown();
-
-        //unsubscribe from "inMemory" broker per topic
-        InMemoryBroker.unsubscribe(subscriberWSO2);
-        InMemoryBroker.unsubscribe(subscriberIBM);
-    }
-
-
-    @Test
-    public void testTextSinkMapperEventGroupDefaultDelimiter() throws InterruptedException {
-        log.info("Test for default delimiter");
-        InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
-            @Override
-            public void onMessage(Object msg) {
-                wso2Count.incrementAndGet();
-            }
-
-            @Override
-            public String getTopic() {
-                return "WSO2";
-            }
-        };
-
-        InMemoryBroker.Subscriber subscriberIBM = new InMemoryBroker.Subscriber() {
-            @Override
-            public void onMessage(Object msg) {
-                ibmCount.incrementAndGet();
-            }
-
-            @Override
-            public String getTopic() {
-                return "IBM";
-            }
-        };
-
-        //subscribe to "inMemory" broker per topic
-        InMemoryBroker.subscribe(subscriberWSO2);
-        InMemoryBroker.subscribe(subscriberIBM);
-
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='true' , " +
-                "@payload(\"Stock price of {{symbol}} is {{price}}\"))) " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        siddhiManager.setExtension("sink:inMemory", InMemorySink.class);
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
-
-        siddhiAppRuntime.start();
-        List<Event> arrayList = new ArrayList<>(10);
-        for (int j = 0; j < 5; j++) {
-            arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"WSO2", 55.6f, 10}));
-            arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"IBM", 75.6f, 10}));
-        }
-        stockStream.send(arrayList.toArray(new io.siddhi.core.event.Event[10]));
-        SiddhiTestHelper.waitForEvents(waitTime, 1, wso2Count, timeout);
-        SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
-
-        //assert event count
-        Assert.assertEquals(wso2Count.get(), 1);
-        Assert.assertEquals(ibmCount.get(), 1);
-        siddhiAppRuntime.shutdown();
-
-        //unsubscribe from "inMemory" broker per topic
-        InMemoryBroker.unsubscribe(subscriberWSO2);
-        InMemoryBroker.unsubscribe(subscriberIBM);
-    }
-
-    @Test
-    public void testTextSinkMapperEventGroupCustomDelimiter() throws InterruptedException {
-        log.info("Test for custom delimiter.");
-        InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
-            @Override
-            public void onMessage(Object msg) {
-                wso2Count.incrementAndGet();
-            }
-
-            @Override
-            public String getTopic() {
-                return "WSO2";
-            }
-        };
-
-        InMemoryBroker.Subscriber subscriberIBM = new InMemoryBroker.Subscriber() {
-            @Override
-            public void onMessage(Object msg) {
-                ibmCount.incrementAndGet();
-            }
-
-            @Override
-            public String getTopic() {
-                return "IBM";
-            }
-        };
-
-        //subscribe to "inMemory" broker per topic
-        InMemoryBroker.subscribe(subscriberWSO2);
-        InMemoryBroker.subscribe(subscriberIBM);
-
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='true'," +
-                "delimiter='#######' , @payload(\"Stock price of {{symbol}} is" +
-                " {{price}}\"))) " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        siddhiManager.setExtension("sink:inMemory", InMemorySink.class);
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
-
-        siddhiAppRuntime.start();
-        List<Event> arrayList = new ArrayList<>(100);
-        for (int j = 0; j < 5; j++) {
-            arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"WSO2", 55.6f, 10}));
-            arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"IBM", 75.6f, 10}));
-        }
-        stockStream.send(arrayList.toArray(new io.siddhi.core.event.Event[10]));
-        SiddhiTestHelper.waitForEvents(waitTime, 1, wso2Count, timeout);
-        SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
-
-        //assert event count
-        Assert.assertEquals(wso2Count.get(), 1);
-        Assert.assertEquals(ibmCount.get(), 1);
-        siddhiAppRuntime.shutdown();
-
-        //unsubscribe from "inMemory" broker per topic
-        InMemoryBroker.unsubscribe(subscriberWSO2);
-        InMemoryBroker.unsubscribe(subscriberIBM);
-    }
-
-    @Test
-    public void testTextSinkMapperWithoutPayload() throws InterruptedException {
-        log.info("Test custom mapping without payload.");
+    public void testTextSinkMapperWithDefaultMapping() throws InterruptedException {
+        log.info("Test default text mapping.");
 
         InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
             @Override
@@ -348,9 +132,9 @@ public class TextCustomSinkMapperTestCase {
         InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
 
         siddhiAppRuntime.start();
-        stockStream.send(new Object[]{"WSO2", 55.6f, 100L});
-        stockStream.send(new Object[]{"IBM", 75.6f, 100L});
-        stockStream.send(new Object[]{"WSO2", 57.6f, 100L});
+        stockStream.send(new Object[] {"WSO2", 55.6f, 100L});
+        stockStream.send(new Object[] {"IBM", 75.6f, 100L});
+        stockStream.send(new Object[] {"WSO2", 57.6f, 100L});
         SiddhiTestHelper.waitForEvents(waitTime, 2, wso2Count, timeout);
         SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
 
@@ -365,8 +149,8 @@ public class TextCustomSinkMapperTestCase {
     }
 
     @Test
-    public void testTextSinkMapperEventGroupSingleEvent() throws InterruptedException {
-        log.info("Test for event group with single event.");
+    public void testTextSinkMapperDefaultEventGrouping() throws InterruptedException {
+        log.info("Test without event grouping attribute.");
         InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
             @Override
             public void onMessage(Object msg) {
@@ -398,8 +182,7 @@ public class TextCustomSinkMapperTestCase {
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
                 "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='true' , " +
-                "@payload(\"Stock price of {{symbol}} is {{price}}\"))) " +
+                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='false')) " +
                 "define stream BarStream (symbol string, price float, volume long); ";
 
         String query = "" +
@@ -413,8 +196,9 @@ public class TextCustomSinkMapperTestCase {
         InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
 
         siddhiAppRuntime.start();
-        stockStream.send(new Object[]{"WSO2", 55.6f, 10});
-        stockStream.send(new Object[]{"IBM", 75.6f, 10});
+
+        stockStream.send(new Object[] {"WSO2", 55.6f, 10});
+        stockStream.send(new Object[] {"IBM", 55.6f, 10});
         SiddhiTestHelper.waitForEvents(waitTime, 1, wso2Count, timeout);
         SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
 
@@ -428,16 +212,13 @@ public class TextCustomSinkMapperTestCase {
         InMemoryBroker.unsubscribe(subscriberIBM);
     }
 
-    @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void testTextSinkWrongMapping() throws InterruptedException {
-        log.info("Test custom for wrong mapping.");
-        List<Object> onMessageList = new ArrayList<Object>();
-
+    @Test
+    public void testTextSinkMapperDefaultEventGroupingTrue() throws InterruptedException {
+        log.info("Test without event grouping attribute.");
         InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
             @Override
             public void onMessage(Object msg) {
                 wso2Count.incrementAndGet();
-                onMessageList.add(msg);
             }
 
             @Override
@@ -450,7 +231,6 @@ public class TextCustomSinkMapperTestCase {
             @Override
             public void onMessage(Object msg) {
                 ibmCount.incrementAndGet();
-                onMessageList.add(msg);
             }
 
             @Override
@@ -466,8 +246,7 @@ public class TextCustomSinkMapperTestCase {
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
                 "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text', @payload(\"Stock price of {{id}} is" +
-                " {{price}}\"))) " +
+                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='true')) " +
                 "define stream BarStream (symbol string, price float, volume long); ";
 
         String query = "" +
@@ -481,16 +260,15 @@ public class TextCustomSinkMapperTestCase {
         InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
 
         siddhiAppRuntime.start();
-        stockStream.send(new Object[]{"WSO2", 55.6f, 100L});
-        stockStream.send(new Object[]{"IBM", 75.6f, 100L});
-        stockStream.send(new Object[]{"WSO2", 57.6f, 100L});
-        SiddhiTestHelper.waitForEvents(waitTime, 0, wso2Count, timeout);
-        SiddhiTestHelper.waitForEvents(waitTime, 0, ibmCount, timeout);
+
+        stockStream.send(new Object[] {"WSO2", 55.6f, 10});
+        stockStream.send(new Object[] {"IBM", 55.6f, 10});
+        SiddhiTestHelper.waitForEvents(waitTime, 1, wso2Count, timeout);
+        SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
 
         //assert event count
-        Assert.assertEquals(wso2Count.get(), 0);
-        Assert.assertEquals(ibmCount.get(), 0);
-
+        Assert.assertEquals(wso2Count.get(), 1);
+        Assert.assertEquals(ibmCount.get(), 1);
         siddhiAppRuntime.shutdown();
 
         //unsubscribe from "inMemory" broker per topic
@@ -499,8 +277,8 @@ public class TextCustomSinkMapperTestCase {
     }
 
     @Test
-    public void testTextSinkMapperNotEventGroupWithCustomDelimiter() throws InterruptedException {
-        log.info("Test extra delimiter.");
+    public void testTextSinkMapperEventGroupDefaultDecimeter() throws InterruptedException {
+        log.info("Test for default event delimiter.");
         InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
             @Override
             public void onMessage(Object msg) {
@@ -532,9 +310,78 @@ public class TextCustomSinkMapperTestCase {
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
                 "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='false'," +
-                "delimiter='#######' , @payload(\"Stock price of {{symbol}} is" +
-                " {{price}}\"))) " +
+                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='true')) " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        siddhiManager.setExtension("sink:inMemory", InMemorySink.class);
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
+
+        siddhiAppRuntime.start();
+        List<io.siddhi.core.event.Event> arrayList = new ArrayList<>(100);
+        for (int j = 0; j < 5; j++) {
+            arrayList.add(new io.siddhi.core.event
+                    .Event(System.currentTimeMillis(), new Object[] {"WSO2", 55.6f, 10}));
+            arrayList.add(new io.siddhi.core.event
+                    .Event(System.currentTimeMillis(), new Object[] {"IBM", 75.6f, 10}));
+        }
+        stockStream.send(arrayList.toArray(new io.siddhi.core.event.Event[10]));
+
+        SiddhiTestHelper.waitForEvents(waitTime, 1, wso2Count, timeout);
+        SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
+
+        //assert event count
+        Assert.assertEquals(wso2Count.get(), 1);
+        Assert.assertEquals(ibmCount.get(), 1);
+        siddhiAppRuntime.shutdown();
+
+        //unsubscribe from "inMemory" broker per topic
+        InMemoryBroker.unsubscribe(subscriberWSO2);
+        InMemoryBroker.unsubscribe(subscriberIBM);
+    }
+
+    @Test
+    public void testTextSinkMapperEventGroupCustomDelimiter() throws InterruptedException {
+        log.info("Test for custom event delimiter.");
+        InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
+            @Override
+            public void onMessage(Object msg) {
+                wso2Count.incrementAndGet();
+            }
+
+            @Override
+            public String getTopic() {
+                return "WSO2";
+            }
+        };
+
+        InMemoryBroker.Subscriber subscriberIBM = new InMemoryBroker.Subscriber() {
+            @Override
+            public void onMessage(Object msg) {
+                ibmCount.incrementAndGet();
+            }
+
+            @Override
+            public String getTopic() {
+                return "IBM";
+            }
+        };
+
+        //subscribe to "inMemory" broker per topic
+        InMemoryBroker.subscribe(subscriberWSO2);
+        InMemoryBroker.subscribe(subscriberIBM);
+
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='true'," +
+                "delimiter='#######')) " +
                 "define stream BarStream (symbol string, price float, volume long); ";
 
         String query = "" +
@@ -551,17 +398,17 @@ public class TextCustomSinkMapperTestCase {
         List<Event> arrayList = new ArrayList<>(100);
         for (int j = 0; j < 5; j++) {
             arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"WSO2", 55.6f, 10}));
+                    .Event(System.currentTimeMillis(), new Object[] {"WSO2", 55.6f, 10}));
             arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"IBM", 75.6f, 10}));
+                    .Event(System.currentTimeMillis(), new Object[] {"IBM", 75.6f, 10}));
         }
         stockStream.send(arrayList.toArray(new io.siddhi.core.event.Event[10]));
-        SiddhiTestHelper.waitForEvents(waitTime, 5, wso2Count, timeout);
-        SiddhiTestHelper.waitForEvents(waitTime, 5, ibmCount, timeout);
+        SiddhiTestHelper.waitForEvents(waitTime, 1, wso2Count, timeout);
+        SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
 
         //assert event count
-        Assert.assertEquals(wso2Count.get(), 5);
-        Assert.assertEquals(ibmCount.get(), 5);
+        Assert.assertEquals(wso2Count.get(), 1);
+        Assert.assertEquals(ibmCount.get(), 1);
         siddhiAppRuntime.shutdown();
 
         //unsubscribe from "inMemory" broker per topic
@@ -571,7 +418,7 @@ public class TextCustomSinkMapperTestCase {
 
     @Test
     public void testTextSinkMapperNewLineCharacter() throws InterruptedException {
-        log.info("Test with custom new lie charater.");
+        log.info("Test for new.line.charater");
         InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
             @Override
             public void onMessage(Object msg) {
@@ -603,9 +450,8 @@ public class TextCustomSinkMapperTestCase {
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
                 "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='false'," +
-                "new.line.character='\\n' , @payload(\"Stock price of {{symbol}} is" +
-                " {{price}}\"))) " +
+                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='false' , new" +
+                ".line.character='\\n')) " +
                 "define stream BarStream (symbol string, price float, volume long); ";
 
         String query = "" +
@@ -619,20 +465,15 @@ public class TextCustomSinkMapperTestCase {
         InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
 
         siddhiAppRuntime.start();
-        List<Event> arrayList = new ArrayList<>(100);
-        for (int j = 0; j < 5; j++) {
-            arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"WSO2", 55.6f, 10}));
-            arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"IBM", 75.6f, 10}));
-        }
-        stockStream.send(arrayList.toArray(new io.siddhi.core.event.Event[10]));
-        SiddhiTestHelper.waitForEvents(waitTime, 5, wso2Count, timeout);
-        SiddhiTestHelper.waitForEvents(waitTime, 5, ibmCount, timeout);
+
+        stockStream.send(new Object[] {"WSO2", 55.6f, 10});
+        stockStream.send(new Object[] {"IBM", 55.6f, 10});
+        SiddhiTestHelper.waitForEvents(waitTime, 1, wso2Count, timeout);
+        SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
 
         //assert event count
-        Assert.assertEquals(wso2Count.get(), 5);
-        Assert.assertEquals(ibmCount.get(), 5);
+        Assert.assertEquals(wso2Count.get(), 1);
+        Assert.assertEquals(ibmCount.get(), 1);
         siddhiAppRuntime.shutdown();
 
         //unsubscribe from "inMemory" broker per topic
@@ -641,8 +482,8 @@ public class TextCustomSinkMapperTestCase {
     }
 
     @Test
-    public void testTextSinkMapperNewLineCharacterWrong() throws InterruptedException {
-        log.info("Test with custom new.line.character.");
+    public void testTextSinkMapperEventGroupSingleEvent() throws InterruptedException {
+        log.info("Test for single event when event grouping enabled.");
         InMemoryBroker.Subscriber subscriberWSO2 = new InMemoryBroker.Subscriber() {
             @Override
             public void onMessage(Object msg) {
@@ -674,9 +515,7 @@ public class TextCustomSinkMapperTestCase {
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
                 "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='false'," +
-                "new.line.character='\\r' , @payload(\"Stock price of {{symbol}} is" +
-                " {{price}}\"))) " +
+                "@sink(type='inMemory', topic='{{symbol}}', @map(type='text' , event.grouping.enabled='true')) " +
                 "define stream BarStream (symbol string, price float, volume long); ";
 
         String query = "" +
@@ -690,20 +529,14 @@ public class TextCustomSinkMapperTestCase {
         InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
 
         siddhiAppRuntime.start();
-        List<Event> arrayList = new ArrayList<>(100);
-        for (int j = 0; j < 5; j++) {
-            arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"WSO2", 55.6f, 10}));
-            arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"IBM", 75.6f, 10}));
-        }
-        stockStream.send(arrayList.toArray(new io.siddhi.core.event.Event[10]));
-        SiddhiTestHelper.waitForEvents(waitTime, 5, wso2Count, timeout);
-        SiddhiTestHelper.waitForEvents(waitTime, 5, ibmCount, timeout);
+        stockStream.send(new Object[] {"WSO2", 55.6f, 10});
+        stockStream.send(new Object[] {"IBM", 75.6f, 10});
+        SiddhiTestHelper.waitForEvents(waitTime, 1, wso2Count, timeout);
+        SiddhiTestHelper.waitForEvents(waitTime, 1, ibmCount, timeout);
 
         //assert event count
-        Assert.assertEquals(wso2Count.get(), 5);
-        Assert.assertEquals(ibmCount.get(), 5);
+        Assert.assertEquals(wso2Count.get(), 1);
+        Assert.assertEquals(ibmCount.get(), 1);
         siddhiAppRuntime.shutdown();
 
         //unsubscribe from "inMemory" broker per topic
@@ -712,8 +545,8 @@ public class TextCustomSinkMapperTestCase {
     }
 
     @Test(enabled = false)
-    public void fileSinkTest() throws InterruptedException {
-        log.info("test text custom map with file io");
+    public void fileSinkTestSingle() throws InterruptedException {
+        log.info("test text default map with file io");
         AtomicInteger count = new AtomicInteger();
         ClassLoader classLoader = TextDefaultSinkMapperTestCase.class.getClassLoader();
         String rootPath = classLoader.getResource("files").getFile();
@@ -721,8 +554,7 @@ public class TextCustomSinkMapperTestCase {
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
                 "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='file', @map(type='text' , @payload('Stock price of {{symbol}} is {{price}}')), " +
-                "append='false', " +
+                "@sink(type='file', @map(type='text'), append='false', " +
                 "file.uri='" + sinkUri + "/{{symbol}}.txt') " +
                 "define stream BarStream (symbol string, price float, volume long); ";
 
@@ -737,10 +569,10 @@ public class TextCustomSinkMapperTestCase {
 
         siddhiAppRuntime.start();
 
-        stockStream.send(new Object[]{"WSO2", 55.6f, 100L});
-        stockStream.send(new Object[]{"IBM", 57.678f, 100L});
-        stockStream.send(new Object[]{"GOOGLE", 50f, 100L});
-        stockStream.send(new Object[]{"REDHAT", 50f, 100L});
+        stockStream.send(new Object[] {"WSO2", 55.6f, 100L});
+        stockStream.send(new Object[] {"IBM", 57.678f, 100L});
+        stockStream.send(new Object[] {"GOOGLE", 50f, 100L});
+        stockStream.send(new Object[] {"REDHAT", 50f, 100L});
         Thread.sleep(100);
 
         List<String> symbolNames = new ArrayList<>();
@@ -781,8 +613,7 @@ public class TextCustomSinkMapperTestCase {
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
                 "define stream FooStream (symbol string, price float, volume long); " +
-                "@sink(type='file', @map(type='text',event.grouping.enabled='true', @payload('Stock price of " +
-                "{{symbol}} is {{price}}')" +
+                "@sink(type='file', @map(type='text',event.grouping.enabled='true'" +
                 "), append='false', " +
                 "file.uri='" + sinkUri + "/{{symbol}}.txt') " +
                 "define stream BarStream (symbol string, price float, volume long); ";
@@ -801,12 +632,12 @@ public class TextCustomSinkMapperTestCase {
         List<Event> arrayListWSO2 = new ArrayList<>(100);
         for (int j = 0; j < 5; j++) {
             arrayListWSO2.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"WSO2", 55.6f, 10}));
+                    .Event(System.currentTimeMillis(), new Object[] {"WSO2", 55.6f, 10}));
         }
         List<Event> arrayListIBM = new ArrayList<>(100);
         for (int j = 0; j < 5; j++) {
             arrayListIBM.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"IBM", 75.6f, 10}));
+                    .Event(System.currentTimeMillis(), new Object[] {"IBM", 75.6f, 10}));
         }
         stockStream.send(arrayListWSO2.toArray(new io.siddhi.core.event.Event[5]));
         stockStream.send(arrayListIBM.toArray(new io.siddhi.core.event.Event[5]));
@@ -815,8 +646,6 @@ public class TextCustomSinkMapperTestCase {
         List<String> symbolNames = new ArrayList<>();
         symbolNames.add("WSO2.txt");
         symbolNames.add("IBM.txt");
-        symbolNames.add("GOOGLE.txt");
-        symbolNames.add("REDHAT.txt");
 
         File sink = new File(sinkUri);
         if (sink.isDirectory()) {
@@ -840,12 +669,73 @@ public class TextCustomSinkMapperTestCase {
         }
     }
 
+    @Test(enabled = false)
+    public void fileSinkTestNewLineCharacter() throws InterruptedException {
+        log.info("test text default map with file io");
+        AtomicInteger count = new AtomicInteger();
+        ClassLoader classLoader = TextDefaultSinkMapperTestCase.class.getClassLoader();
+        String rootPath = classLoader.getResource("files").getFile();
+        String sinkUri = rootPath + "/sink";
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "@sink(type='file', @map(type='text',event.grouping.enabled='true'," +
+                "delimiter='#######',new" +
+                ".line.character='\\r\\n'), append='false', " +
+                "file.uri='" + sinkUri + "/{{symbol}}.txt') " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        InputHandler stockStream = siddhiAppRuntime.getInputHandler("FooStream");
+
+        siddhiAppRuntime.start();
+
+        stockStream.send(new Object[] {"WSO2", 55.6f, 100L});
+        stockStream.send(new Object[] {"IBM", 57.678f, 100L});
+        stockStream.send(new Object[] {"GOOGLE", 50f, 100L});
+        stockStream.send(new Object[] {"REDHAT", 50f, 100L});
+        Thread.sleep(100);
+
+        List<String> symbolNames = new ArrayList<>();
+        symbolNames.add("WSO2.txt");
+        symbolNames.add("IBM.txt");
+        symbolNames.add("GOOGLE.txt");
+        symbolNames.add("REDHAT.txt");
+
+        File sink = new File(sinkUri);
+        if (sink.isDirectory()) {
+            for (File file : sink.listFiles()) {
+                if (symbolNames.contains(file.getName())) {
+                    count.incrementAndGet();
+                }
+            }
+            AssertJUnit.assertEquals(4, count.intValue());
+        } else {
+            AssertJUnit.fail(sinkUri + " is not a directory.");
+        }
+
+        Thread.sleep(1000);
+        siddhiAppRuntime.shutdown();
+        File sinkRoot = new File(sinkUri);
+        try {
+            FileUtils.deleteDirectory(sinkRoot);
+        } catch (IOException e) {
+            throw new TestException("Failed to delete files in '" + sinkUri + "' due to " + e.getMessage(), e);
+        }
+    }
+
     /**
      * Creating test for publishing events with TEXT mapping.
      *
      * @throws Exception Interrupted exception
      */
-    @Test(enabled =  false)
+    @Test(enabled = false)
     public void testHTTPTextMappingText() throws Exception {
 
         log.info("Creating test for publishing events with TEXT mapping.");
@@ -854,7 +744,7 @@ public class TextCustomSinkMapperTestCase {
         String inStreamDefinition = "Define stream FooStream (message String,method String,headers String);"
                 + "@sink(type='http',publisher.url='http://localhost:8005/abc',method='{{method}}',"
                 + "headers='{{headers}}',"
-                + "@map(type='text', @payload('{{{message}}}'))) "
+                + "@map(type='text')) "
                 + "Define stream BarStream (message String,method String,headers String);";
         String query = ("@info(name = 'query1') " +
                 "from FooStream select message,method,headers insert into BarStream;");
@@ -864,57 +754,70 @@ public class TextCustomSinkMapperTestCase {
         siddhiAppRuntime.start();
         HttpServerListenerHandler listenerHandler = new HttpServerListenerHandler(8005);
         listenerHandler.run();
-        fooStream.send(new Object[]{"WSO2,55.6,100", "POST", "'Name:John','Age:23'"});
+        fooStream.send(new Object[] {"WSO2,55.6,100", "POST", "'Name:John','Age:23'"});
         while (!listenerHandler.getServerListener().isMessageArrive()) {
             Thread.sleep(10);
         }
         String eventData = listenerHandler.getServerListener().getData();
-        Assert.assertEquals(eventData, "WSO2,55.6,100\n");
+        Assert.assertEquals(eventData, "message:\"WSO2,55.6,100\",\n" +
+                "method:\"POST\",\n" +
+                "headers:\"'Name:John','Age:23'\"\n");
         listenerHandler.shutdown();
         siddhiAppRuntime.shutdown();
     }
 
     @Test(enabled = false)
-    public void testHTTPTextMappingTextMultiple() throws Exception {
+    public void testTextSinkMapperEventGroupHTTP() throws InterruptedException {
+        log.info("Test for default event delimiter and http.");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "define stream FooStream (message String,method String,headers String);" +
+                "@sink(type='http',publisher.url='http://localhost:8005/abc',method='{{method}}',"
+                + "headers='{{headers}}', @map(type='text' , event.grouping" +
+                ".enabled='true')) " +
+                "define stream BarStream (message String,method String,headers String);";
 
-        log.info("Creating test for publishing events with TEXT mapping.");
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
         SiddhiManager siddhiManager = new SiddhiManager();
-        siddhiManager.setExtension("text-output-mapper", TextSinkMapper.class);
-        String inStreamDefinition = "Define stream FooStream (message String,method String,headers String);"
-                + "@sink(type='http',publisher.url='http://localhost:8005/abc',method='{{method}}',"
-                + "headers='{{headers}}',"
-                + "@map(type='text', event.grouping" +
-                ".enabled='true', @payload('{{{message}}}'))) "
-                + "Define stream BarStream (message String,method String,headers String);";
-        String query = ("@info(name = 'query1') " +
-                "from FooStream select message,method,headers insert into BarStream;");
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(inStreamDefinition +
-                query);
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
         InputHandler fooStream = siddhiAppRuntime.getInputHandler("FooStream");
         siddhiAppRuntime.start();
         HttpServerListenerHandler listenerHandler = new HttpServerListenerHandler(8005);
         listenerHandler.run();
-        List<Event> arrayList = new ArrayList<>(5);
+        List<io.siddhi.core.event.Event> arrayList = new ArrayList<>(5);
         for (int j = 0; j < 5; j++) {
             arrayList.add(new io.siddhi.core.event
-                    .Event(System.currentTimeMillis(), new Object[]{"WSO2", "POST", "'place:office'"}));
+                    .Event(System.currentTimeMillis(), new Object[] {"WSO2", "POST", "'place:office'"}));
         }
         fooStream.send(arrayList.toArray(new io.siddhi.core.event.Event[5]));
         while (!listenerHandler.getServerListener().isMessageArrive()) {
             Thread.sleep(10);
         }
         String eventData = listenerHandler.getServerListener().getData();
-        Assert.assertEquals(eventData, "WSO2\n" +
+        Assert.assertEquals(eventData, "message:\"WSO2\",\n" +
+                "method:\"POST\",\n" +
+                "headers:\"'place:office'\"\n" +
                 "~~~~~~~~~~\n" +
-                "WSO2\n" +
+                "message:\"WSO2\",\n" +
+                "method:\"POST\",\n" +
+                "headers:\"'place:office'\"\n" +
                 "~~~~~~~~~~\n" +
-                "WSO2\n" +
+                "message:\"WSO2\",\n" +
+                "method:\"POST\",\n" +
+                "headers:\"'place:office'\"\n" +
                 "~~~~~~~~~~\n" +
-                "WSO2\n" +
+                "message:\"WSO2\",\n" +
+                "method:\"POST\",\n" +
+                "headers:\"'place:office'\"\n" +
                 "~~~~~~~~~~\n" +
-                "WSO2\n");
+                "message:\"WSO2\",\n" +
+                "method:\"POST\",\n" +
+                "headers:\"'place:office'\"\n");
         listenerHandler.shutdown();
         siddhiAppRuntime.shutdown();
     }
-
 }
