+++ component/src/test/java/io/siddhi/extension/map/text/sinkmapper/util/HttpServerListener.java
@@ -16,7 +16,7 @@
  *  under the License.
  *
  */
-package org.wso2.extension.siddhi.map.text.sinkmapper.util;
+package io.siddhi.extension.map.text.sinkmapper.util;
 
 import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpExchange;
