+++ component/src/test/java/io/siddhi/extension/map/text/sourcemapper/TextDefaultSourceMapperTestCase.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.map.text.sourcemapper;
+package io.siddhi.extension.map.text.sourcemapper;
 
 import io.siddhi.core.SiddhiAppRuntime;
 import io.siddhi.core.SiddhiManager;
@@ -28,11 +28,11 @@ import io.siddhi.core.util.SiddhiTestHelper;
 import io.siddhi.core.util.persistence.InMemoryPersistenceStore;
 import io.siddhi.core.util.persistence.PersistenceStore;
 import io.siddhi.core.util.transport.InMemoryBroker;
+import io.siddhi.extension.map.text.sourcemapper.util.HttpTestUtil;
 import org.apache.log4j.Logger;
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import org.wso2.extension.siddhi.map.text.sourcemapper.util.HttpTestUtil;
 
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
