+++ pom.xml
@@ -26,7 +26,7 @@
     <modelVersion>4.0.0</modelVersion>
     <packaging>pom</packaging>
 
-    <groupId>org.wso2.extension.siddhi.map.text</groupId>
+    <groupId>io.siddhi.extension.map.text</groupId>
     <artifactId>siddhi-map-text-parent</artifactId>
     <version>2.0.1-SNAPSHOT</version>
     <name>Siddhi Text Map Extensions</name>
